backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Database
psycopg[binary]==3.1.16
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
greenlet==3.0.3

# Redis & Task Queue
redis==5.0.1
celery==5.3.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography>=41.0.7  # Security fix for CVE-2023-50782

# HTTP & External APIs
httpx==0.26.0
aiohttp>=3.9.3  # Security fixes for CVE-2024-23334, CVE-2024-23829
idna>=3.6  # Security fix for CVE-2024-3651
requests>=2.31.0  # For OAuth and external API calls

# AI/LLM
openai==1.7.2

# SMS (Twilio)
twilio==8.11.0

# Calendar Integration
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
msal==1.26.0

# Validation & Serialization
email-validator==2.1.0.post1

# Templates & Email
jinja2>=3.1.3  # Template engine for email/reports
markupsafe>=2.1.3  # Security for template rendering

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-mock==0.15.0

# Development
ruff==0.1.14
mypy==1.8.0

# Utilities
python-dotenv==1.0.0
tzdata==2023.4

# Security constraints (transitive deps)
certifi>=2023.7.22  # CVE-2023-37920
urllib3>=2.0.7,<3  # Multiple CVEs
jinja2>=3.1.3  # CVE-2024-22195
tornado>=6.3.3  # CVE-2023-28370 (via jupyter/celery)
